+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/config/AuthClientConfiguration.java
@@ -1,0 +1,28 @@
+package ee.cyber.cdoc2.config;
+
+import java.security.KeyStore;
+import java.util.Properties;
+
+import ee.cyber.cdoc2.exceptions.ConfigurationLoadingException;
+
+public interface AuthClientConfiguration {
+
+    static AuthClientConfiguration load(Properties properties)
+        throws ConfigurationLoadingException {
+        return AuthClientConfigurationProps.load(properties);
+    }
+
+    String getHostUrl();
+
+    KeyStore getTrustStore();
+
+    int getReadTimeout();
+
+    int getConnectTimeout();
+
+    int getPollingIntervalMs();
+
+    int getPollingMaxCount();
+
+    boolean getClientServerDebug();
+}