+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/config/ConfigurationProperties.java
@@ -4,9 +4,9 @@ package ee.cyber.cdoc2.config;
 /**
  * CDOC2 configuration properties
  */
-public final class Cdoc2ConfigurationProperties {
+public final class ConfigurationProperties {
 
-    private Cdoc2ConfigurationProperties() {
+    private ConfigurationProperties() {
     }
 
     /**
@@ -67,6 +67,10 @@ public final class Cdoc2ConfigurationProperties {
         = "cdoc2.key-shares.client.ssl.trust-store";
     public static final String KEY_SHARES_CLIENT_TRUST_STORE_PWD
         = "cdoc2.key-shares.client.ssl.trust-store-password";
+    public static final String KEY_SHARES_CLIENT_READ_TIMEOUT
+        = "key-shares.client.read-timeout";
+    public static final String KEY_SHARES_CLIENT_CONNECT_TIMEOUT
+        = "key-shares.client.connect-timeout";
     public static final String KEY_SHARES_CLIENT_TRUST_STORE_TYPE
         = "cdoc2.key-shares.client.ssl.trust-store.type";
 
@@ -108,6 +112,13 @@ public final class Cdoc2ConfigurationProperties {
         "auth-server.client.ssl.trust-store";
     public static final String AUTH_SERVER_CLIENT_TRUST_STORE_PWD
         = "auth-server.client.ssl.trust-store-password";
+    public static final String AUTH_SERVER_CLIENT_READ_TIMEOUT = "auth-server.client.read-timeout";
+    public static final String AUTH_SERVER_CLIENT_CONNECT_TIMEOUT =
+        "auth-server.client.connect-timeout";
+    public static final String AUTH_SERVER_CLIENT_POLLING_INTERVAL_MS =
+        "auth-server.client.polling.intervalMs";
+    public static final String AUTH_SERVER_CLIENT_POLLING_MAX_COUNT =
+        "auth-server.client.polling.maxCount";
 
     /**
      * Defines CDOC2 RP Server properties file location
@@ -118,6 +129,9 @@ public final class Cdoc2ConfigurationProperties {
     public static final String RP_SERVER_CLIENT_TRUST_STORE = "rp-server.client.ssl.trust-store";
     public static final String RP_SERVER_CLIENT_TRUST_STORE_PWD
         = "rp-server.client.ssl.trust-store-password";
+    public static final String RP_SERVER_CLIENT_READ_TIMEOUT = "rp-server.client.read-timeout";
+    public static final String RP_SERVER_CLIENT_CONNECT_TIMEOUT =
+        "rp-server.client.connect-timeout";
 
     public static final String RP_SERVER_MID_DISPLAY_TEXT_FORMAT
         = "rp-server.mid.display-text-format";