+++ cdoc2-lib/src/test/java/ee/cyber/cdoc2/services/Cdoc2ServicesTest.java
@@ -4,7 +4,7 @@ import java.security.GeneralSecurityException;
 
 import org.junit.jupiter.api.Test;
 
-import ee.cyber.cdoc2.client.rpserver.Cdoc2RpClient;
+import ee.cyber.cdoc2.client.RpClient;
 
 import static ee.cyber.cdoc2.ClientConfigurationUtil.DEMO_ENV_PROPERTIES;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -14,6 +14,6 @@ class Cdoc2ServicesTest {
     @Test
     void testInitFromProperties() throws GeneralSecurityException {
         Services services = Cdoc2Services.initFromProperties(DEMO_ENV_PROPERTIES);
-        assertTrue(services.hasService(Cdoc2RpClient.class));
+        assertTrue(services.hasService(RpClient.class));
     }
 }